+++ src/components/Home/Daily.tsx
@@ -10,21 +10,18 @@ import { useState, useEffect } from "react";
 import { useTailwind } from "tailwind-rn/dist";
 import { useDispatch, useSelector } from "react-redux";
 import { AppDispatch } from "@stores/index";
-import { RootReducer } from "../../App";
+import { RootReducer } from "../../../App";
 import {
-  getMonthlyHistories,
+  getDailyHistories,
   History,
   deleteHistory,
   changeMeansuringTime,
-  updatePrimary,
 } from "@stores/history";
 import { VictoryPie } from "victory-native";
 import DropDownPicker from "react-native-dropdown-picker";
 import { dateFormat } from "@utils/format";
-import ChangeInfo from "./ChangeInfo";
+import ChangeInfo from "../ChangeInfo";
 import Modal from "react-native-modal";
-import { Picker } from "@react-native-picker/picker";
-import { PrimaryCategory, getPrimaries } from "@stores/categories";
 
 type item = {
   label: string;
@@ -43,12 +40,11 @@ export default () => {
   const {
     user: { user },
     history: {
-      histories: { monthly },
+      histories: { daily },
     },
-    categories: { primaryCategories },
   } = useSelector((store: RootReducer) => store);
 
-  const monthlyMap = monthly.reduce(
+  const dailyMap = daily.reduce(
     (
       prev: {
         [key: string]: {
@@ -78,11 +74,10 @@ export default () => {
     {}
   );
 
-  const [modalType, setModalType] = useState<
-    "initial" | "editTime" | "updatePrimary" | null
-  >(null);
-  const [selectPrimary, setSelectPrimary] = useState();
-  const [pieData, setPieData] = useState<pie[]>(Object.values(monthlyMap));
+  const [modalType, setModalType] = useState<"initial" | "editTime" | null>(
+    null
+  );
+  const [pieData, setPieData] = useState<pie[]>(Object.values(dailyMap));
   const [open, setOpen] = useState(false);
   const [primary, setPrimary] = useState(null);
   const [primaries, setPrimaries] = useState<item[]>([]);
@@ -90,22 +85,16 @@ export default () => {
   const [secondary, setSecondary] = useState(null);
   const [secondaries, setSecondaries] = useState<item[]>([]);
 
-  useEffect(() => {
-    dispatch(getPrimaries({ userID: user!.id }));
-  }, [user]);
-
   const close = () => setModalType(null);
 
   const InitialModal = ({
     close,
     editTime,
     deleteItem,
-    showPrimary,
   }: {
     close: () => void;
     editTime: () => void;
     deleteItem: () => void;
-    showPrimary: () => void;
   }) => {
     const tailwind = useTailwind();
     return (
@@ -114,7 +103,7 @@ export default () => {
           <Text style={tailwind("text-center text-base")}>
             データの修正はこちらから
           </Text>
-          <Button title="カテゴリ情報の修正" onPress={showPrimary} />
+          <Button title="カテゴリ情報の修正" onPress={close} />
           <Button title="計測時間の修正" onPress={editTime} />
           <Button
             title="データの削除"
@@ -127,20 +116,8 @@ export default () => {
       </>
     );
   };
-  const ShowPrimaryPicker = ({ id }: { id: string }) => {
-    return (
-      <Picker
-        selectedValue={selectPrimary}
-        onValueChange={(item) => setSelectPrimary(item)}
-      >
-        {primaryCategories.map((data) => (
-          <Picker.Item label={data.name} value={data.name} />
-        ))}
-      </Picker>
-    );
-  };
 
-  const EditTimeModal = ({ id, close }: { id: string; close: () => void }) => {
+  const EditTimeModal = ({ id }: { id: string }) => {
     const [time, setTime] = useState("");
     const tailwind = useTailwind();
     return (
@@ -171,19 +148,19 @@ export default () => {
   };
 
   useEffect(() => {
-    if (user) dispatch(getMonthlyHistories({ userId: user!.id }));
+    if (user) dispatch(getDailyHistories({ userId: user!.id }));
   }, [user]);
   useEffect(() => {
-    const primaryInfo = Object.values(monthlyMap).map((item) => {
+    const primaryInfo = Object.values(dailyMap).map((item) => {
       return { label: item.x, value: item.id };
     });
     setPrimaries([...primaryInfo, { label: "全て", value: "all" }]);
-    setPieData(Object.values(monthlyMap));
-  }, [monthly]);
+    setPieData(Object.values(dailyMap));
+  }, [daily]);
 
   useEffect(() => {
     if (primary && primary !== "all") {
-      const secondaryMap = monthlyMap[primary].secondaries.reduce(
+      const secondaryMap = dailyMap[primary].secondaries.reduce(
         (
           pre: {
             [key: string]: {
@@ -225,8 +202,8 @@ export default () => {
     }
 
     if (primary === "all") {
-      setPieData(Object.values(monthlyMap));
-      const newSecondaries = monthly
+      setPieData(Object.values(dailyMap));
+      const newSecondaries = daily
         .filter(
           (x, i, array) =>
             array.findIndex((y) => y.secondary_id === x.secondary_id) === i
@@ -268,15 +245,9 @@ export default () => {
                   deleteItem={() =>
                     dispatch(deleteHistory({ historyId: item.id }))
                   }
-                  showPrimary={() => setModalType("updatePrimary")}
                 />
               )}
-              {modalType == "editTime" && (
-                <EditTimeModal id={item.id} close={close} />
-              )}
-              {modalType == "updatePrimary" && (
-                <ShowPrimaryPicker id={item.id} />
-              )}
+              {modalType == "editTime" && <EditTimeModal id={item.id} />}
             </Modal>
           </View>
           <Text style={tailwind("text-base font-bold text-right mr-1 pr-1")}>
@@ -327,15 +298,13 @@ export default () => {
         />
       </View>
       <FlatList
-        data={monthly}
+        data={daily}
         renderItem={renderItem}
         keyExtractor={(item) => item.id}
         refreshControl={
           <RefreshControl
             refreshing={false}
-            onRefresh={() =>
-              dispatch(getMonthlyHistories({ userId: user!.id }))
-            }
+            onRefresh={() => dispatch(getDailyHistories({ userId: user!.id }))}
           />
         }
       />
