+++ tailwind.css
@@ -22,12 +22,12 @@
   margin: 0.25rem
 }
 
-.m-4 {
-  margin: 1rem
-}
-
 .m-2 {
   margin: 0.5rem
+}
+
+.m-4 {
+  margin: 1rem
 }
 
 .mx-3\.5 {
@@ -43,14 +43,6 @@
 .my-4 {
   margin-top: 1rem;
   margin-bottom: 1rem
-}
-
-.mt-1 {
-  margin-top: 0.25rem
-}
-
-.ml-2 {
-  margin-left: 0.5rem
 }
 
 .mr-1 {
@@ -77,6 +69,18 @@
   margin-top: 2rem
 }
 
+.mt-1 {
+  margin-top: 0.25rem
+}
+
+.ml-2 {
+  margin-left: 0.5rem
+}
+
+.mb-2 {
+  margin-bottom: 0.5rem
+}
+
 .flex {
   display: flex
 }
@@ -101,16 +105,16 @@
   height: 2rem
 }
 
+.w-full {
+  width: 100%
+}
+
 .w-4\/5 {
   width: 80%
 }
 
 .w-1\/2 {
   width: 50%
-}
-
-.w-full {
-  width: 100%
 }
 
 .w-20 {
@@ -145,10 +149,6 @@
   justify-content: center
 }
 
-.rounded-2xl {
-  border-radius: 1rem
-}
-
 .rounded-md {
   border-radius: 0.375rem
 }
@@ -157,8 +157,20 @@
   border-radius: 9999px
 }
 
+.rounded-2xl {
+  border-radius: 1rem
+}
+
 .rounded-lg {
   border-radius: 0.5rem
+}
+
+.rounded {
+  border-radius: 0.25rem
+}
+
+.rounded-xl {
+  border-radius: 0.75rem
 }
 
 .rounded-t-3xl {
@@ -187,9 +199,9 @@
   border-color: rgb(0 0 0 / var(--tw-border-opacity))
 }
 
-.border-violet-400 {
+.border-violet-300 {
   --tw-border-opacity: 1;
-  border-color: rgb(167 139 250 / var(--tw-border-opacity))
+  border-color: rgb(196 181 253 / var(--tw-border-opacity))
 }
 
 .border-white {
@@ -197,24 +209,9 @@
   border-color: rgb(255 255 255 / var(--tw-border-opacity))
 }
 
-.border-violet-900 {
-  --tw-border-opacity: 1;
-  border-color: rgb(76 29 149 / var(--tw-border-opacity))
-}
-
-.border-violet-500 {
-  --tw-border-opacity: 1;
-  border-color: rgb(139 92 246 / var(--tw-border-opacity))
-}
-
-.border-violet-300 {
-  --tw-border-opacity: 1;
-  border-color: rgb(196 181 253 / var(--tw-border-opacity))
-}
-
-.bg-white {
+.bg-violet-500 {
   --tw-bg-opacity: 1;
-  background-color: rgb(255 255 255 / var(--tw-bg-opacity))
+  background-color: rgb(139 92 246 / var(--tw-bg-opacity))
 }
 
 .bg-yellow-200 {
@@ -222,14 +219,14 @@
   background-color: rgb(254 240 138 / var(--tw-bg-opacity))
 }
 
-.bg-teal-500 {
+.bg-black {
   --tw-bg-opacity: 1;
-  background-color: rgb(20 184 166 / var(--tw-bg-opacity))
+  background-color: rgb(0 0 0 / var(--tw-bg-opacity))
 }
 
-.bg-red-500 {
+.bg-white {
   --tw-bg-opacity: 1;
-  background-color: rgb(239 68 68 / var(--tw-bg-opacity))
+  background-color: rgb(255 255 255 / var(--tw-bg-opacity))
 }
 
 .bg-sky-400 {
@@ -237,45 +234,20 @@
   background-color: rgb(56 189 248 / var(--tw-bg-opacity))
 }
 
-.bg-black {
-  --tw-bg-opacity: 1;
-  background-color: rgb(0 0 0 / var(--tw-bg-opacity))
-}
-
-.bg-violet-500 {
-  --tw-bg-opacity: 1;
-  background-color: rgb(139 92 246 / var(--tw-bg-opacity))
-}
-
-.bg-violet-400 {
-  --tw-bg-opacity: 1;
-  background-color: rgb(167 139 250 / var(--tw-bg-opacity))
-}
-
 .p-2 {
   padding: 0.5rem
-}
-
-.p-12 {
-  padding: 3rem
 }
 
 .p-1 {
   padding: 0.25rem
 }
 
+.p-12 {
+  padding: 3rem
+}
+
 .p-4 {
   padding: 1rem
-}
-
-.px-2 {
-  padding-left: 0.5rem;
-  padding-right: 0.5rem
-}
-
-.py-1 {
-  padding-top: 0.25rem;
-  padding-bottom: 0.25rem
 }
 
 .px-4 {
@@ -288,8 +260,14 @@
   padding-bottom: 1rem
 }
 
-.pl-1 {
-  padding-left: 0.25rem
+.px-2 {
+  padding-left: 0.5rem;
+  padding-right: 0.5rem
+}
+
+.py-1 {
+  padding-top: 0.25rem;
+  padding-bottom: 0.25rem
 }
 
 .pr-1 {
@@ -300,6 +278,14 @@
   padding-top: 0.5rem
 }
 
+.pl-1 {
+  padding-left: 0.25rem
+}
+
+.pt-12 {
+  padding-top: 3rem
+}
+
 .text-center {
   text-align: center
 }
@@ -308,19 +294,29 @@
   text-align: right
 }
 
-.text-base {
-  font-size: 1rem;
-  line-height: 1.5rem
-}
-
 .text-2xl {
   font-size: 1.5rem;
   line-height: 2rem
 }
 
+.text-base {
+  font-size: 1rem;
+  line-height: 1.5rem
+}
+
 .text-4xl {
   font-size: 2.25rem;
   line-height: 2.5rem
+}
+
+.text-lg {
+  font-size: 1.125rem;
+  line-height: 1.75rem
+}
+
+.text-xl {
+  font-size: 1.25rem;
+  line-height: 1.75rem
 }
 
 .font-bold {
@@ -330,11 +326,6 @@
 .text-white {
   --tw-text-opacity: 1;
   color: rgb(255 255 255 / var(--tw-text-opacity))
-}
-
-.text-violet-400 {
-  --tw-text-opacity: 1;
-  color: rgb(167 139 250 / var(--tw-text-opacity))
 }
 
 .text-violet-500 {
