+++ tailwind.json
@@ -32,20 +32,20 @@
 			"marginLeft": 4
 		}
 	},
-	"m-4": {
-		"style": {
-			"marginTop": 16,
-			"marginRight": 16,
-			"marginBottom": 16,
-			"marginLeft": 16
-		}
-	},
 	"m-2": {
 		"style": {
 			"marginTop": 8,
 			"marginRight": 8,
 			"marginBottom": 8,
 			"marginLeft": 8
+		}
+	},
+	"m-4": {
+		"style": {
+			"marginTop": 16,
+			"marginRight": 16,
+			"marginBottom": 16,
+			"marginLeft": 16
 		}
 	},
 	"mx-3.5": {
@@ -64,16 +64,6 @@
 		"style": {
 			"marginTop": 16,
 			"marginBottom": 16
-		}
-	},
-	"mt-1": {
-		"style": {
-			"marginTop": 4
-		}
-	},
-	"ml-2": {
-		"style": {
-			"marginLeft": 8
 		}
 	},
 	"mr-1": {
@@ -106,6 +96,21 @@
 			"marginTop": 32
 		}
 	},
+	"mt-1": {
+		"style": {
+			"marginTop": 4
+		}
+	},
+	"ml-2": {
+		"style": {
+			"marginLeft": 8
+		}
+	},
+	"mb-2": {
+		"style": {
+			"marginBottom": 8
+		}
+	},
 	"flex": {
 		"style": {
 			"display": "flex"
@@ -136,6 +141,11 @@
 			"height": 32
 		}
 	},
+	"w-full": {
+		"style": {
+			"width": "100%"
+		}
+	},
 	"w-4/5": {
 		"style": {
 			"width": "80%"
@@ -144,11 +154,6 @@
 	"w-1/2": {
 		"style": {
 			"width": "50%"
-		}
-	},
-	"w-full": {
-		"style": {
-			"width": "100%"
 		}
 	},
 	"w-20": {
@@ -193,14 +198,6 @@
 			"justifyContent": "center"
 		}
 	},
-	"rounded-2xl": {
-		"style": {
-			"borderTopLeftRadius": 16,
-			"borderTopRightRadius": 16,
-			"borderBottomRightRadius": 16,
-			"borderBottomLeftRadius": 16
-		}
-	},
 	"rounded-md": {
 		"style": {
 			"borderTopLeftRadius": 6,
@@ -217,12 +214,36 @@
 			"borderBottomLeftRadius": 9999
 		}
 	},
+	"rounded-2xl": {
+		"style": {
+			"borderTopLeftRadius": 16,
+			"borderTopRightRadius": 16,
+			"borderBottomRightRadius": 16,
+			"borderBottomLeftRadius": 16
+		}
+	},
 	"rounded-lg": {
 		"style": {
 			"borderTopLeftRadius": 8,
 			"borderTopRightRadius": 8,
 			"borderBottomRightRadius": 8,
 			"borderBottomLeftRadius": 8
+		}
+	},
+	"rounded": {
+		"style": {
+			"borderTopLeftRadius": 4,
+			"borderTopRightRadius": 4,
+			"borderBottomRightRadius": 4,
+			"borderBottomLeftRadius": 4
+		}
+	},
+	"rounded-xl": {
+		"style": {
+			"borderTopLeftRadius": 12,
+			"borderTopRightRadius": 12,
+			"borderBottomRightRadius": 12,
+			"borderBottomLeftRadius": 12
 		}
 	},
 	"rounded-t-3xl": {
@@ -269,13 +290,13 @@
 			"borderLeftColor": "rgb(0 0 0 / var(--tw-border-opacity))"
 		}
 	},
-	"border-violet-400": {
+	"border-violet-300": {
 		"style": {
 			"--tw-border-opacity": 1,
-			"borderTopColor": "rgb(167 139 250 / var(--tw-border-opacity))",
-			"borderRightColor": "rgb(167 139 250 / var(--tw-border-opacity))",
-			"borderBottomColor": "rgb(167 139 250 / var(--tw-border-opacity))",
-			"borderLeftColor": "rgb(167 139 250 / var(--tw-border-opacity))"
+			"borderTopColor": "rgb(196 181 253 / var(--tw-border-opacity))",
+			"borderRightColor": "rgb(196 181 253 / var(--tw-border-opacity))",
+			"borderBottomColor": "rgb(196 181 253 / var(--tw-border-opacity))",
+			"borderLeftColor": "rgb(196 181 253 / var(--tw-border-opacity))"
 		}
 	},
 	"border-white": {
@@ -287,37 +308,10 @@
 			"borderLeftColor": "rgb(255 255 255 / var(--tw-border-opacity))"
 		}
 	},
-	"border-violet-900": {
-		"style": {
-			"--tw-border-opacity": 1,
-			"borderTopColor": "rgb(76 29 149 / var(--tw-border-opacity))",
-			"borderRightColor": "rgb(76 29 149 / var(--tw-border-opacity))",
-			"borderBottomColor": "rgb(76 29 149 / var(--tw-border-opacity))",
-			"borderLeftColor": "rgb(76 29 149 / var(--tw-border-opacity))"
-		}
-	},
-	"border-violet-500": {
-		"style": {
-			"--tw-border-opacity": 1,
-			"borderTopColor": "rgb(139 92 246 / var(--tw-border-opacity))",
-			"borderRightColor": "rgb(139 92 246 / var(--tw-border-opacity))",
-			"borderBottomColor": "rgb(139 92 246 / var(--tw-border-opacity))",
-			"borderLeftColor": "rgb(139 92 246 / var(--tw-border-opacity))"
-		}
-	},
-	"border-violet-300": {
-		"style": {
-			"--tw-border-opacity": 1,
-			"borderTopColor": "rgb(196 181 253 / var(--tw-border-opacity))",
-			"borderRightColor": "rgb(196 181 253 / var(--tw-border-opacity))",
-			"borderBottomColor": "rgb(196 181 253 / var(--tw-border-opacity))",
-			"borderLeftColor": "rgb(196 181 253 / var(--tw-border-opacity))"
-		}
-	},
-	"bg-white": {
+	"bg-violet-500": {
 		"style": {
 			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(255 255 255 / var(--tw-bg-opacity))"
+			"backgroundColor": "rgb(139 92 246 / var(--tw-bg-opacity))"
 		}
 	},
 	"bg-yellow-200": {
@@ -326,40 +320,22 @@
 			"backgroundColor": "rgb(254 240 138 / var(--tw-bg-opacity))"
 		}
 	},
-	"bg-teal-500": {
-		"style": {
-			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(20 184 166 / var(--tw-bg-opacity))"
-		}
-	},
-	"bg-red-500": {
-		"style": {
-			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(239 68 68 / var(--tw-bg-opacity))"
-		}
-	},
-	"bg-sky-400": {
-		"style": {
-			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(56 189 248 / var(--tw-bg-opacity))"
-		}
-	},
 	"bg-black": {
 		"style": {
 			"--tw-bg-opacity": 1,
 			"backgroundColor": "rgb(0 0 0 / var(--tw-bg-opacity))"
 		}
 	},
-	"bg-violet-500": {
+	"bg-white": {
 		"style": {
 			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(139 92 246 / var(--tw-bg-opacity))"
+			"backgroundColor": "rgb(255 255 255 / var(--tw-bg-opacity))"
 		}
 	},
-	"bg-violet-400": {
+	"bg-sky-400": {
 		"style": {
 			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(167 139 250 / var(--tw-bg-opacity))"
+			"backgroundColor": "rgb(56 189 248 / var(--tw-bg-opacity))"
 		}
 	},
 	"p-2": {
@@ -370,14 +346,6 @@
 			"paddingLeft": 8
 		}
 	},
-	"p-12": {
-		"style": {
-			"paddingTop": 48,
-			"paddingRight": 48,
-			"paddingBottom": 48,
-			"paddingLeft": 48
-		}
-	},
 	"p-1": {
 		"style": {
 			"paddingTop": 4,
@@ -386,24 +354,20 @@
 			"paddingLeft": 4
 		}
 	},
+	"p-12": {
+		"style": {
+			"paddingTop": 48,
+			"paddingRight": 48,
+			"paddingBottom": 48,
+			"paddingLeft": 48
+		}
+	},
 	"p-4": {
 		"style": {
 			"paddingTop": 16,
 			"paddingRight": 16,
 			"paddingBottom": 16,
 			"paddingLeft": 16
-		}
-	},
-	"px-2": {
-		"style": {
-			"paddingLeft": 8,
-			"paddingRight": 8
-		}
-	},
-	"py-1": {
-		"style": {
-			"paddingTop": 4,
-			"paddingBottom": 4
 		}
 	},
 	"px-4": {
@@ -418,9 +382,16 @@
 			"paddingBottom": 16
 		}
 	},
-	"pl-1": {
+	"px-2": {
 		"style": {
-			"paddingLeft": 4
+			"paddingLeft": 8,
+			"paddingRight": 8
+		}
+	},
+	"py-1": {
+		"style": {
+			"paddingTop": 4,
+			"paddingBottom": 4
 		}
 	},
 	"pr-1": {
@@ -433,6 +404,16 @@
 			"paddingTop": 8
 		}
 	},
+	"pl-1": {
+		"style": {
+			"paddingLeft": 4
+		}
+	},
+	"pt-12": {
+		"style": {
+			"paddingTop": 48
+		}
+	},
 	"text-center": {
 		"style": {
 			"textAlign": "center"
@@ -443,22 +424,34 @@
 			"textAlign": "right"
 		}
 	},
-	"text-base": {
-		"style": {
-			"fontSize": 16,
-			"lineHeight": 24
-		}
-	},
 	"text-2xl": {
 		"style": {
 			"fontSize": 24,
 			"lineHeight": 32
 		}
 	},
+	"text-base": {
+		"style": {
+			"fontSize": 16,
+			"lineHeight": 24
+		}
+	},
 	"text-4xl": {
 		"style": {
 			"fontSize": 36,
 			"lineHeight": 40
+		}
+	},
+	"text-lg": {
+		"style": {
+			"fontSize": 18,
+			"lineHeight": 28
+		}
+	},
+	"text-xl": {
+		"style": {
+			"fontSize": 20,
+			"lineHeight": 28
 		}
 	},
 	"font-bold": {
@@ -470,12 +463,6 @@
 		"style": {
 			"--tw-text-opacity": 1,
 			"color": "rgb(255 255 255 / var(--tw-text-opacity))"
-		}
-	},
-	"text-violet-400": {
-		"style": {
-			"--tw-text-opacity": 1,
-			"color": "rgb(167 139 250 / var(--tw-text-opacity))"
 		}
 	},
 	"text-violet-500": {
